// types.ts
export type TypographyModifier =
  | 'Page'
  | 'H1'
  | 'H2'
  | 'H3'
  | 'H4'
  | 'H5'
  | 'H6'
  | 'TitleSection'
  | 'TitleBlockMedium'
  | 'TitleBlockSmall'
  | 'TitleBlockSmallUpper'
  | 'FormFieldLabel'
  | 'TextNote'
  | 'ItemPriceBefore'
  | 'ItemPriceCurrent'
  | 'ItemCommission'
  | 'ScreenTitle'
  | 'ScreenSubTitle'
  | 'Text3xLgBold'
  | 'Text2xLgBold'
  | 'TextXLgBold'
  | 'TextLgBold'
  | 'TextMdBold'
  | 'TextNmBold'
  | 'TextSmBold'
  | 'TextXsBold'
  | 'Text2XsBold'
  | 'Text3xLgMedium'
  | 'Text2xLgMedium'
  | 'TextXLgMedium'
  | 'TextLgMedium'
  | 'TextMdMedium'
  | 'TextNmMedium'
  | 'TextSmMedium'
  | 'TextXsMedium'
  | 'Text2XsMedium'
  | 'Text3xLgNormal'
  | 'Text2xLgNormal'
  | 'TextXLgNormal'
  | 'TextLgNormal'
  | 'TextMdNormal'
  | 'TextNmNormal'
  | 'TextSmNormal'
  | 'TextXsNormal'
  | 'Text2XsNormal'
  | 'Display3xLgBold'
  | 'Display2xLgBold'
  | 'DisplayXLgBold'
  | 'DisplayLgBold'
  | 'DisplayMdBold'
  | 'DisplayNmBold'
  | 'DisplaySmBold'
  | 'DisplayXsBold'
  | 'Display2XsBold'
  | 'Display3xLgMedium'
  | 'Display2xLgMedium'
  | 'DisplayXLgMedium'
  | 'DisplayLgMedium'
  | 'DisplayMdMedium'
  | 'DisplayNmMedium'
  | 'DisplaySmMedium'
  | 'DisplayXsMedium'
  | 'Display2XsMedium'
  | 'Display3xLgNormal'
  | 'Display2xLgNormal'
  | 'DisplayXLgNormal'
  | 'DisplayLgNormal'
  | 'DisplayMdNormal'
  | 'DisplayNmNormal'
  | 'DisplaySmNormal'
  | 'DisplayXsNormal'
  | 'Display2XsNormal';
// ...
export interface TestModifier {
  testID?: string;
}
// ...
export type KAvatarSize = 'xs' | 'sm' | 'md' | 'lg' | 'xlg' | '2xlg' | '3xlg' | '4xlg';
// ...
export interface KAvatarProps
  extends React.PropsWithChildren<{
      uri: string | number;
      size?: KAvatarSize;
      showOnlineDot?: boolean;
      isOnline?: boolean;
      placeholder?: number;
      tintColor?: string;
      border?: {
        width: number;
        color?: string;
      };
      onPress?: () => void;
    }>,
    TestModifier {}
// ...
export type KPopupActionProps = {
  label: string;
  weight?: KButtonLabelWeight;
  icon?: KButtonIconProps;
  onPress?: () => void;
  kind?: KButtonKind;
};

export type KPopupActionsProps = {
  type?:
    | 'vertical-menu'
    | 'inline'
    | 'single'
    | 'horizontal-menu'
    | 'primary-action'
    | 'horizontal-button';
  buttons: Array<KPopupActionProps>;
  primaryIndex?: number; // the highest priority
  destructiveIndex?: number; // the normal priority
};

export interface KPopupProps {
  body: {
    message?: {
      text: string;
      typo?: TypographyModifier;
      color?: string;
    };
    renderContent?: (dismiss: () => void) => any;
    scrollable?: boolean;
  };
  header?: {
    title?: {
      text: string;
      typo?: TypographyModifier;
      color?: string;
    };
    subTitle?: {
      text: string;
      typo?: TypographyModifier;
      color?: string;
    };
    alignment?: 'center' | 'left';
    logo?: KAvatarProps;
    showCloseButton?: boolean;
  };
  actions?: KPopupActionsProps;
  position?: 'top' | 'center' | 'bottom';
  touchOutsideToDismiss?: boolean;
  backgroundColor?: string;
}
// ...
export interface WithPopupProps {
  open: (payload: KPopupProps) => void;
  dismiss: () => void;
  dismissAll: () => void;
}